Refactor persons entity

+++ src/main/java/ru/deloom/jpaservice/repositories/PersonsRepository.java
@@ -5,6 +5,6 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.deloom.jpaservice.entities.Persons;
 
-public interface PersonsRepository extends JpaRepository<Persons, Integer>{
+public interface PersonsRepository extends JpaRepository<Persons, Long>{
 	List<Persons> findByName(String name);
 }
